daemon: update dfget command

+++ client/daemon/peer/peertask_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestBitmap_Sets(t *testing.T) {
 	b := NewBitmap()
-	for i := 0; i < 512; i++ {
+	for i := int32(0); i < 512; i++ {
 		b.Set(i)
 	}
 	b.Sets(2, 3, 3, 4)
